Added an IIdentityProviderConfigurationService so we can retrieve all the configured 3rd party identifiers.
