+++ Fabric.Identity.API/Quickstart/Account/AccountService.cs
@@ -10,6 +10,7 @@ using Microsoft.AspNetCore.Http;
 using System.Linq;
 using System.Threading.Tasks;
 using Fabric.Identity.API.Configuration;
+using Fabric.Identity.API.Services;
 
 namespace IdentityServer4.Quickstart.UI
 {
@@ -19,17 +20,20 @@ namespace IdentityServer4.Quickstart.UI
         private readonly IIdentityServerInteractionService _interaction;
         private readonly IHttpContextAccessor _httpContextAccessor;
         private readonly IAppConfiguration _appConfiguration;
+        private readonly IIdentityProviderConfigurationService _identityProviderConfigurationService;
 
         public AccountService(
             IIdentityServerInteractionService interaction,
             IHttpContextAccessor httpContextAccessor,
             IClientStore clientStore,
-            IAppConfiguration appConfiguration)
+            IAppConfiguration appConfiguration,
+            IIdentityProviderConfigurationService identityProviderConfigurationService)
         {
             _interaction = interaction;
             _httpContextAccessor = httpContextAccessor;
             _clientStore = clientStore;
             _appConfiguration = appConfiguration;
+            _identityProviderConfigurationService = identityProviderConfigurationService;
         }
 
         public async Task<LoginViewModel> BuildLoginViewModelAsync(string returnUrl)
@@ -47,29 +51,7 @@ namespace IdentityServer4.Quickstart.UI
                 };
             }
 
-            var schemes = _httpContextAccessor.HttpContext.Authentication.GetAuthenticationSchemes();
-
-            var providers = schemes
-                .Where(x => x.DisplayName != null && !AccountOptions.WindowsAuthenticationSchemes.Contains(x.AuthenticationScheme))
-                .Select(x => new ExternalProvider
-                {
-                    DisplayName = x.DisplayName,
-                    AuthenticationScheme = x.AuthenticationScheme
-                }).ToList();
-
-            if (AccountOptions.WindowsAuthenticationEnabled)
-            {
-                // this is needed to handle windows auth schemes
-                var windowsSchemes = schemes.Where(s => AccountOptions.WindowsAuthenticationSchemes.Contains(s.AuthenticationScheme));
-                if (windowsSchemes.Any())
-                {
-                    providers.Add(new ExternalProvider
-                    {
-                        AuthenticationScheme = AccountOptions.WindowsAuthenticationSchemes.First(),
-                        DisplayName = AccountOptions.WindowsAuthenticationDisplayName
-                    });
-                }
-            }
+            var providers = _identityProviderConfigurationService.GetConfiguredIdentityProviders();
 
             var allowLocal = true;
             if (context?.ClientId != null)
